Fix nasty indentation bug in circuit thread.
router.guard was never set unless the router was itself a guard,
causing test circuits to not be build, but the router's last_tested
field would be set anyway, making it believe it was already tested.
This error propagated in quite a nasty fashion, making the testing
cycle appear much shorter (~65 nodes instead of ~500).