Implement test retry handling.
We now attempt to retry failed circuits three times before giving up on
the node for one test cycle (Hammer).  We also give up early if:
  - The consensus says the router is hibernating
  - The router has been out of the consensus for torbel_config.stale_router_timeout
    seconds.
  - The router is flagged as a BadExit.

This is very helpful as initial testing shows that around 50% of retries
are successful.
On second thought, since the retry success rate is so high, it might be
detrimental to give up early on certain routers.  More on this later.