New test data implementation.
Slightly different than before: send the router's test circuit ID
in addition to their ID.  This makes it much easier to discard potentially
old test connections that the router took a large amount of time to complete.
This also makes it very difficult for a router to try and mess with TorBEL,
since they have to know the current circuit ID assigned to it, which is
difficult to guess.