Fix is_tor_traffic; rename to tor_exit_search, change API.
Works now for routers that haven't been tested in an export.
tor_exit_search better reflects what the call actually does,
which searches for and returns information about a potential
router on the Tor network.

tor_exit_search now requires the target IP address, which it
should have done in the first place.