Fix deadlock condition in ConservativeScheduler.
For real this time.  Caused by acquiring locks out-of-order;
it only occurred during circuit failures that happened at
(nearly) the same time as a fetch_next_tests call.  Tested
and tested and tested again, ran for a very long time with
no deadlock, so I assume it is safe.  Will test more later.