Implement torbel/tordnsel consistency checker.
Currently checks the current consensus against a smattering
of random IP addresses and random ports as well as the ports
torbel is configured to test.  We report on whether TorBEL,
the public TorDNSEL, and a local TorBEL-powered DNSEL server
agree on exit status.

TODO: Test more sources than just the current consensus.